Improved handling of ChatML tokens, im_end and im_start in LLM generated response  (#64)
Problem:

With the Microsoft Phi-3 model and JSON processing we noticed that the generated response from the LLM (sometimes) includes multiple occurrences of the Chat ML tokens, “im_start” and “im_end”. New hallucinated questions are also added to the response. In certain cases these newly introduced questions (by the LLM) are relevant to the end-user’s question. In some instances, these are not related.

Solution in this PR:

Currently this will be processed as follows:
* Any content after (and including) the first im_end ChatML token and the keywords "Question:", "Content:" and "Context:" is trimmed.
* Within the remaining content, we replace any occurrence of im_start with a space.