Update Chat UI to display history of question and answers (#63)
Currently the QA in a box UI only shows the current question typed in by the user and corresponding answer generated by the LLM. Prospective customers wanted to be able to view prior questions, so they could use context from the generated answer in a follow up questions.

This PR, updates the Chat UI to show the history of all questions asked by the user during the current session.

This is implemented by setting chat history to be enabled by default.

Note that this updates only the display. Prior questions and answers are not sent to the LLM as context for a subsequent question.