Add a question router to choose between SQL and vector (hybrid) search (#71)
This PR includes a question router. This router determines whether an incoming question will be handled by the SQL search or the RAG vector search. RAG vector search is parametrized to enable hybrid search i.e. a combination of vector search and text search.

The question classification module uses a random forest classifier trained against synthetic questions (not included in this repo). The two classes are
-     "aggregation" questions
-     "pointed" questions.

SQL search code path is chosen for:
A. all aggregation questions, as well as,
B. questions that have symbols or numerals in the questions.

B is included in the SQL search type, because vector search does not handle alphanumeric words in a question such as: IP addresses, ticket numbers, etc based on our experiments.

Squashed commits: 

* Random forest model for question classification

* Add question router

* Return unique sources

* Update file paths for SQL DB and classification models to containerized value

---------

Co-authored-by: Maciej Urbański <maciek@elotl.co>